refactor(web-app): move global components to the Home page (#126)
close #126
