+++ apps/web-app/src/components/feature/index.ts
@@ -1,7 +1,3 @@
-export * from './AddWindowsDropdownMenu'
 export * from './AppLogo'
-export * from './CompatibilityWindow/CompatibilityWindow'
 export * from './FormTextField/FormTextField'
-export * from './RecordingWindow/RecordingWindow'
-export * from './WatchRecordingWindow/WatchRecordingWindow'
