+++ apps/web-app/src/layouts/index.ts
@@ -1,4 +1,3 @@
 export * from './DividedLayout/DividedLayout'
-export * from './MainLayout'
 export * from './SectionLayout/SectionLayout'
