+++ apps/web-app/src/pages/Home/components/feature/CompatibilityWindow/CompatibilityWindow.tsx
