+++ apps/web-app/src/pages/Home/components/feature/CompatibilityWindow/hooks/index.ts
