+++ apps/web-app/src/pages/Home/components/feature/CompatibilityWindow/models/default-values.model.ts
