+++ apps/web-app/src/pages/Home/components/feature/CompatibilityWindow/models/index.ts
