+++ apps/web-app/src/pages/Home/components/feature/RecordingWindow/RecordingWindow.tsx
