+++ apps/web-app/src/pages/Home/components/feature/RecordingWindow/components/index.ts
