+++ apps/web-app/src/pages/Home/components/feature/index.ts
@@ -1,0 +1,6 @@
+export * from './AddWindowsDropdownMenu'
+export * from './CompatibilityWindow/CompatibilityWindow'
+export * from './DownloadRecordingWindow/DownloadRecordingWindow'
+export * from './RecordingWindow/RecordingWindow'
+export * from './WatchRecordingWindow/WatchRecordingWindow'
+
