+++ apps/web-app/src/pages/Home/layouts/MainLayout.tsx
@@ -1,11 +1,12 @@
-import { AddWindowsDropdownMenu, AppLogo, CompatibilityWindow } from '@/components/feature'
+import { AppLogo } from '@/components/feature'
 import { type BaseComponentType } from '@/models'
+import { AddWindowsDropdownMenu, CompatibilityWindow } from '@/pages/Home/components/feature'
 
 export const MainLayout: BaseComponentType = ({ children }) => {
   return (
     <div className='flex flex-col [height:100svh] [font-family:Roboto] [font-weight:500]'>
       <CompatibilityWindow />
-      <AppLogo/>
+      <AppLogo />
       <AddWindowsDropdownMenu />
       <main className='h-full text-sm'>{children}</main>
     </div>
