+++ apps/web-app/src/pages/Home/layouts/index.ts
@@ -1,0 +1,1 @@
+export * from './MainLayout'
